Move prepare_sst_for_ingestion to sst_importer crate and handle encryption
