+++ components/raftstore/src/errors.rs
@@ -11,8 +11,6 @@ use prost::{DecodeError, EncodeError};
 use protobuf::ProtobufError;
 
 use kvproto::{errorpb, metapb};
-use pd_client;
-use raft;
 use tikv_util::codec;
 
 use super::coprocessor::Error as CopError;
@@ -149,6 +147,12 @@ quick_error! {
             cause(err)
             description(err.description())
             display("Snapshot {}", err)
+        }
+        SstImporter(err: sst_importer::Error) {
+            from()
+            cause(err)
+            description(err.description())
+            display("SstImporter {}", err)
         }
     }
 }
