+++ components/raftstore/src/store/snap.rs
@@ -715,12 +715,11 @@ impl Snap {
             check_file_size_and_checksum(&cf_file.path, cf_file.size, cf_file.checksum, mgr)?;
 
             if !for_send && !plain_file_used(cf_file.cf) {
-                engine.prepare_sst_for_ingestion(&cf_file.path, &cf_file.clone_path)?;
-                if let Some(mgr) = self.encryption_key_manager.as_ref() {
-                    let src = cf_file.path.to_str().unwrap();
-                    let dst = cf_file.clone_path.to_str().unwrap();
-                    mgr.link_file(src, dst)?;
-                }
+                sst_importer::prepare_sst_for_ingestion(
+                    &cf_file.path,
+                    &cf_file.clone_path,
+                    &self.encryption_key_manager,
+                )?;
             }
         }
         Ok(())
