+++ components/sst_importer/src/errors.rs
@@ -6,6 +6,7 @@ use std::num::ParseIntError;
 use std::path::PathBuf;
 use std::result;
 
+use encryption::Error as EncryptionError;
 use grpcio::Error as GrpcError;
 use kvproto::import_sstpb;
 use tokio_sync::oneshot::error::RecvError;
@@ -28,6 +29,7 @@ pub fn error_inc(err: &Error) {
         Error::CannotReadExternalStorage(..) => "read_external_storage",
         Error::WrongKeyPrefix(..) => "wrong_prefix",
         Error::BadFormat(..) => "bad_format",
+        Error::Encryption(..) => "encryption",
         _ => return,
     };
     IMPORTER_ERROR_VEC.with_label_values(&[label]).inc();
@@ -97,6 +99,11 @@ quick_error! {
         }
         BadFormat(msg: String) {
             display("bad format {}", msg)
+        }
+        Encryption(err: EncryptionError) {
+            from()
+            description("encryption error")
+            display("Encryption {:?}", err)
         }
     }
 }
